Fix attn_res precision: only softmax in fp32, cast weights back
Match Flash Attention's precision pattern: softmax computed in fp32 for
numerical stability, weights cast back to input dtype so the weighted
sum einsum stays in ambient precision (bf16/fp16) instead of pulling
everything to fp32 via type promotion.

+++ nano/model.py
@@ -95,8 +95,8 @@ def attn_res(states, proj, norm):
     K = norm(V)
     logits = torch.einsum('d, n b t d -> n b t', proj.weight.squeeze(), K)
     with torch.amp.autocast(V.device.type, enabled=False):
-        weights = logits.float().softmax(0)
-        h = torch.einsum('n b t, n b t d -> b t d', weights, V.float()).to(V.dtype)
+        weights = logits.float().softmax(0).to(logits.dtype)
+    h = torch.einsum('n b t, n b t d -> b t d', weights, V)
     return h
